# docs/swagger.yaml
definitions:
  app.Pager:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_rows:
        type: integer
    type: object
  errcode.Error:
    type: object
  model.Link:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      exp:
        type: integer
      id:
        type: integer
      originURL:
        description: 原始链接
        type: string
      sc:
        description: 短链代码
        type: string
      sid:
        description: 业务标识
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  model.LinkSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Link'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
        type: object
    type: object
  service.CreateLinkRequest:
    properties:
      expiration_in_minutes:
        type: integer
      url:
        type: string
    required:
    - url
    type: object
info:
  contact: {}
  description: GoGo
  license: {}
  termsOfService: mm
  title: 短链接服务
  version: "1.0"
paths:
  /a/{sc}:
    get:
      parameters:
      - description: 短链接ID
        in: path
        name: sc
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 短链接跳转
  /api/v1/link-trace:
    get:
      parameters:
      - description: 状态
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.LinkSwagger'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个短链接访问记录
  /api/v1/link-trace/{sc}:
    get:
      parameters:
      - description: 短链sc
        in: path
        name: sc
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Link'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取短链的访问记录
  /api/v1/links:
    get:
      parameters:
      - description: 创建者ID
        in: query
        name: created_by
        type: integer
      - description: 状态
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.LinkSwagger'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个短链接
    post:
      consumes:
      - application/json
      parameters:
      - description: 链接信息
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/service.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功，短链数据
          schema:
            $ref: '#/definitions/model.Link'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新增短链接
  /api/v1/links/{sc}:
    get:
      parameters:
      - description: 短链ID
        in: path
        name: sc
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Link'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取单个短链详情
swagger: "2.0"